alert 이상하게 나오는 버그 수정 트레이너 별점 insert 처리 select 해야됨

+++ OrderSystem/src/main/java/kr/ac/kopo/model/EatPlanner.java
@@ -8,7 +8,7 @@ public class EatPlanner {
 	private String eatName;
 	private int eatGram;
 	private int eatCount;
-	private int eatKcal;
+	private float eatKcal;
 	private String eatEtc;
 	private String eatNutrient;
 	private String eatTime;
@@ -24,11 +24,12 @@ public class EatPlanner {
 		this.eatGram = eatGram;
 	}
 
-	public int getEatKcal() {
+	
+	public float getEatKcal() {
 		return eatKcal;
 	}
 
-	public void setEatKcal(int eatKcal) {
+	public void setEatKcal(float eatKcal) {
 		this.eatKcal = eatKcal;
 	}
